src/test/java/com/xjtu/dao/UserDaoTest.java
package com.xjtu.dao;

import com.xjtu.domain.User;
import org.junit.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.test.context.ContextConfiguration;
import org.springframework.test.context.junit4.AbstractTransactionalJUnit4SpringContextTests;
import org.springframework.test.context.transaction.TransactionConfiguration;

import static org.junit.Assert.assertEquals;

@ContextConfiguration(locations = {"classpath:applicationContext.xml"})
@TransactionConfiguration(defaultRollback = false)
public class UserDaoTest extends AbstractTransactionalJUnit4SpringContextTests {

    @Autowired
    private UserDao userDao;

    @Test
    public void should_select_all_names() throws Exception {
        userDao.deleteAll();
        User user = new User("xjty", "wangyan");
        userDao.save(user);
        assertEquals(1, userDao.all().size());
    }

}
